Move external task process to the main example project and remove example-external-task-process which is no longer needed.
