--- example-external-task-process/build.gradle
@@ -1,49 +1,0 @@
-plugins {
-    id("com.github.johnrengelman.shadow")
-    id("io.micronaut.application")
-
-    // used for tests
-    id("org.jetbrains.kotlin.jvm")
-    id("org.jetbrains.kotlin.kapt")
-    id("org.jetbrains.kotlin.plugin.allopen")
-}
-
-group = "info.novatec"
-
-micronaut {
-    runtime("jetty")
-    testRuntime("junit5")
-    processing {
-        incremental(true)
-        annotations("micronaut.camunda.bpm.externaltask.process.*")
-    }
-}
-
-dependencies {
-    implementation(project(":micronaut-camunda-bpm-feature"))
-    implementation("io.micronaut:micronaut-runtime")
-    // needed to initialize FetchAndLockContextListener
-    implementation("org.camunda.bpm:camunda-engine-rest-jaxrs2:$camundaVersion")
-
-    // Dependency is currently needed if the build is executed with the Netty runtime in the CI's matrix build
-    compileOnly("javax.servlet:javax.servlet-api:4.0.1")
-
-    runtimeOnly("ch.qos.logback:logback-classic")
-    runtimeOnly("com.h2database:h2")
-
-    kaptTest platform("io.micronaut:micronaut-bom:$micronautVersion")
-    kaptTest("io.micronaut.data:micronaut-data-processor")
-    kapt("io.micronaut:micronaut-inject-java:$micronautVersion")
-
-    testImplementation("org.camunda.bpm.assert:camunda-bpm-assert:8.0.0")
-    testImplementation("org.assertj:assertj-core")
-}
-
-application {
-    mainClass.set("info.novatec.micronaut.camunda.externaltask.process.ExternalTaskProcessApplication")
-}
-
-java {
-    sourceCompatibility = JavaVersion.toVersion("1.8")
-    targetCompatibility = JavaVersion.toVersion("1.8")
-}
