--- example-external-task-process/src/main/resources/external-task-process.bpmn
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" id="Definitions_1cv4zb0" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="4.5.0">
-  <bpmn:process id="externalTaskProcessExample" isExecutable="true" camunda:versionTag="1.0.0">
-    <bpmn:sequenceFlow id="SequenceFlow_0zvob4z" sourceRef="StartEvent" targetRef="createItem" />
-    <bpmn:endEvent id="EndEvent" name="Processed data">
-      <bpmn:incoming>Flow_1sgpapd</bpmn:incoming>
-    </bpmn:endEvent>
-    <bpmn:sequenceFlow id="SequenceFlow_17q8g0e" sourceRef="createItem" targetRef="externalTask" />
-    <bpmn:startEvent id="StartEvent" name="Repeat every 5s with 99 repetitions">
-      <bpmn:outgoing>SequenceFlow_0zvob4z</bpmn:outgoing>
-      <bpmn:timerEventDefinition>
-        <bpmn:timeCycle xsi:type="bpmn:tFormalExpression">R99/PT5S</bpmn:timeCycle>
-      </bpmn:timerEventDefinition>
-    </bpmn:startEvent>
-    <bpmn:serviceTask id="createItem" name="create Item" camunda:delegateExpression="#{itemDelegate}">
-      <bpmn:incoming>SequenceFlow_0zvob4z</bpmn:incoming>
-      <bpmn:outgoing>SequenceFlow_17q8g0e</bpmn:outgoing>
-    </bpmn:serviceTask>
-    <bpmn:sequenceFlow id="SequenceFlow_0eg8114" sourceRef="externalTask" targetRef="logResult" />
-    <bpmn:serviceTask id="externalTask" name="External Task" camunda:type="external" camunda:topic="my-topic">
-      <bpmn:incoming>SequenceFlow_17q8g0e</bpmn:incoming>
-      <bpmn:outgoing>SequenceFlow_0eg8114</bpmn:outgoing>
-    </bpmn:serviceTask>
-    <bpmn:sequenceFlow id="Flow_1sgpapd" sourceRef="logResult" targetRef="EndEvent" />
-    <bpmn:serviceTask id="logResult" name="log result" camunda:delegateExpression="#{resultDelegate}">
-      <bpmn:incoming>SequenceFlow_0eg8114</bpmn:incoming>
-      <bpmn:outgoing>Flow_1sgpapd</bpmn:outgoing>
-    </bpmn:serviceTask>
-  </bpmn:process>
-  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
-    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="externalTaskProcessExample">
-      <bpmndi:BPMNEdge id="Flow_1sgpapd_di" bpmnElement="Flow_1sgpapd">
-        <di:waypoint x="740" y="121" />
-        <di:waypoint x="792" y="121" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="SequenceFlow_0eg8114_di" bpmnElement="SequenceFlow_0eg8114">
-        <di:waypoint x="560" y="121" />
-        <di:waypoint x="640" y="121" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="SequenceFlow_17q8g0e_di" bpmnElement="SequenceFlow_17q8g0e">
-        <di:waypoint x="370" y="121" />
-        <di:waypoint x="460" y="121" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="SequenceFlow_0zvob4z_di" bpmnElement="SequenceFlow_0zvob4z">
-        <di:waypoint x="215" y="121" />
-        <di:waypoint x="270" y="121" />
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNShape id="EndEvent_0p65cti_di" bpmnElement="EndEvent">
-        <dc:Bounds x="792" y="103" width="36" height="36" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="771" y="146" width="78" height="14" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="StartEvent_0390bbf_di" bpmnElement="StartEvent">
-        <dc:Bounds x="179" y="103" width="36" height="36" />
-        <bpmndi:BPMNLabel>
-          <dc:Bounds x="153" y="146" width="89" height="27" />
-        </bpmndi:BPMNLabel>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="ServiceTask_18metwe_di" bpmnElement="createItem">
-        <dc:Bounds x="270" y="81" width="100" height="80" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="ServiceTask_1a0g76n_di" bpmnElement="externalTask">
-        <dc:Bounds x="460" y="81" width="100" height="80" />
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Activity_19txgdf_di" bpmnElement="logResult">
-        <dc:Bounds x="640" y="81" width="100" height="80" />
-      </bpmndi:BPMNShape>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-</bpmn:definitions>
