+++ micronaut-camunda-bpm-example/src/main/java/info/novatec/micronaut/camunda/bpm/example/calculation/PrintResultDelegate.java
@@ -13,30 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package info.novatec.micronaut.camunda.externaltask.process;
+package info.novatec.micronaut.camunda.bpm.example.calculation;
 
-import java.util.concurrent.ThreadLocalRandom;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Singleton;
-import java.util.concurrent.atomic.AtomicInteger;
 
 @Singleton
-public class ItemDelegate implements JavaDelegate {
+public class PrintResultDelegate implements JavaDelegate {
 
-    private static final Logger log = LoggerFactory.getLogger(ItemDelegate.class);
-
-    private final AtomicInteger lastCount = new AtomicInteger(0);
+    private static final Logger log = LoggerFactory.getLogger(PrintResultDelegate.class);
 
     @Override
-    public void execute(DelegateExecution delegateExecution) {
-        int itemValue = ThreadLocalRandom.current().nextInt(0, 1001);
-        delegateExecution.setVariable("itemValue", itemValue);
-        int itemNumber = lastCount.incrementAndGet();
-        delegateExecution.setVariable("itemNumber", itemNumber);
-        log.debug("Process Instance for item {} with Value {} running.", itemNumber, itemValue);
+    public void execute(DelegateExecution execution) {
+        log.info("Process instance for item {} with value {} is approved={}",
+                execution.getVariable("itemNumber"),
+                execution.getVariable("itemValue"),
+                execution.getVariable("approved"));
     }
 }
