+++ micronaut-camunda-bpm-example/src/main/resources/calculation.bpmn
@@ -1,0 +1,109 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:camunda="http://camunda.org/schema/1.0/bpmn" id="Definitions_1hzovk7" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Camunda Modeler" exporterVersion="4.6.0">
+  <bpmn:process id="Calculation" isExecutable="true">
+    <bpmn:sequenceFlow id="SequenceFlow_0fxbboj" sourceRef="StartEvent_1" targetRef="ServiceTask_CreateItem" />
+    <bpmn:endEvent id="EndEvent_03gbeuf" name="Completed">
+      <bpmn:incoming>Flow_1p2h80x</bpmn:incoming>
+    </bpmn:endEvent>
+    <bpmn:sequenceFlow id="SequenceFlow_1uip4un" sourceRef="ServiceTask_CreateItem" targetRef="ServiceTask_Calculate" />
+    <bpmn:startEvent id="StartEvent_1" name="Every Second">
+      <bpmn:outgoing>SequenceFlow_0fxbboj</bpmn:outgoing>
+      <bpmn:timerEventDefinition>
+        <bpmn:timeCycle xsi:type="bpmn:tFormalExpression">* * * * * ?</bpmn:timeCycle>
+      </bpmn:timerEventDefinition>
+    </bpmn:startEvent>
+    <bpmn:serviceTask id="ServiceTask_CreateItem" name="Create Item" camunda:delegateExpression="${createItemDelegate}">
+      <bpmn:incoming>SequenceFlow_0fxbboj</bpmn:incoming>
+      <bpmn:outgoing>SequenceFlow_1uip4un</bpmn:outgoing>
+    </bpmn:serviceTask>
+    <bpmn:sequenceFlow id="Flow_0kvupiu" sourceRef="ServiceTask_Calculate" targetRef="ServiceTask_PrintResult" />
+    <bpmn:serviceTask id="ServiceTask_Calculate" name="Do complicated calculation" camunda:type="external" camunda:topic="my-topic">
+      <bpmn:incoming>SequenceFlow_1uip4un</bpmn:incoming>
+      <bpmn:outgoing>Flow_0kvupiu</bpmn:outgoing>
+    </bpmn:serviceTask>
+    <bpmn:boundaryEvent id="TimerEvent_Wait" name="Ten seconds" attachedToRef="ServiceTask_Calculate">
+      <bpmn:outgoing>Flow_09cr6sc</bpmn:outgoing>
+      <bpmn:timerEventDefinition id="TimerEventDefinition_1twdpcq">
+        <bpmn:timeDuration xsi:type="bpmn:tFormalExpression">PT10S</bpmn:timeDuration>
+      </bpmn:timerEventDefinition>
+    </bpmn:boundaryEvent>
+    <bpmn:endEvent id="Event_1tdykkt" name="Cancelled">
+      <bpmn:incoming>Flow_09cr6sc</bpmn:incoming>
+    </bpmn:endEvent>
+    <bpmn:sequenceFlow id="Flow_09cr6sc" sourceRef="TimerEvent_Wait" targetRef="Event_1tdykkt" />
+    <bpmn:serviceTask id="ServiceTask_PrintResult" name="Print Result" camunda:delegateExpression="${printResultDelegate}">
+      <bpmn:incoming>Flow_0kvupiu</bpmn:incoming>
+      <bpmn:outgoing>Flow_1p2h80x</bpmn:outgoing>
+    </bpmn:serviceTask>
+    <bpmn:sequenceFlow id="Flow_1p2h80x" sourceRef="ServiceTask_PrintResult" targetRef="EndEvent_03gbeuf" />
+    <bpmn:textAnnotation id="TextAnnotation_0h50m82">
+      <bpmn:text>This makes sure that we don't have more than two instances running if the job executor runs every five seconds (default)</bpmn:text>
+    </bpmn:textAnnotation>
+    <bpmn:association id="Association_079qo1o" sourceRef="TimerEvent_Wait" targetRef="TextAnnotation_0h50m82" />
+  </bpmn:process>
+  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="Calculation">
+      <bpmndi:BPMNEdge id="Flow_09cr6sc_di" bpmnElement="Flow_09cr6sc">
+        <di:waypoint x="460" y="178" />
+        <di:waypoint x="460" y="240" />
+        <di:waypoint x="712" y="240" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="Flow_0kvupiu_di" bpmnElement="Flow_0kvupiu">
+        <di:waypoint x="490" y="120" />
+        <di:waypoint x="550" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="SequenceFlow_1uip4un_di" bpmnElement="SequenceFlow_1uip4un">
+        <di:waypoint x="359" y="120" />
+        <di:waypoint x="390" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="SequenceFlow_0fxbboj_di" bpmnElement="SequenceFlow_0fxbboj">
+        <di:waypoint x="209" y="120" />
+        <di:waypoint x="259" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="Flow_1p2h80x_di" bpmnElement="Flow_1p2h80x">
+        <di:waypoint x="650" y="120" />
+        <di:waypoint x="712" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNShape id="EndEvent_03gbeuf_di" bpmnElement="EndEvent_03gbeuf">
+        <dc:Bounds x="712" y="102" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="703" y="78" width="54" height="14" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="StartEvent_15l3yd7_di" bpmnElement="StartEvent_1">
+        <dc:Bounds x="173" y="102" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="157" y="145" width="69" height="14" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="ServiceTask_0dc28o5_di" bpmnElement="ServiceTask_CreateItem">
+        <dc:Bounds x="259" y="80" width="100" height="80" />
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="Activity_0prdlco_di" bpmnElement="ServiceTask_Calculate">
+        <dc:Bounds x="390" y="80" width="100" height="80" />
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="Event_1tdykkt_di" bpmnElement="Event_1tdykkt">
+        <dc:Bounds x="712" y="222" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="706" y="265" width="49" height="14" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="TextAnnotation_0h50m82_di" bpmnElement="TextAnnotation_0h50m82">
+        <dc:Bounds x="350" y="270" width="250" height="54" />
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="Activity_14uqcm9_di" bpmnElement="ServiceTask_PrintResult">
+        <dc:Bounds x="550" y="80" width="100" height="80" />
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="Event_0nbmgx3_di" bpmnElement="TimerEvent_Wait">
+        <dc:Bounds x="442" y="142" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="479" y="183" width="62" height="14" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge id="Association_079qo1o_di" bpmnElement="Association_079qo1o">
+        <di:waypoint x="451" y="175" />
+        <di:waypoint x="397" y="270" />
+      </bpmndi:BPMNEdge>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+</bpmn:definitions>
