+++ micronaut-camunda-bpm-example/src/test/java/info/novatec/micronaut/camunda/bpm/example/CamundaControllerTest.java
@@ -45,6 +45,6 @@ class CamundaControllerTest {
         HttpRequest<String> request = HttpRequest.GET("/definitions");
         String body = client.toBlocking().retrieve(request);
 
-        assertEquals("HelloWorld,Onboarding", body);
+        assertEquals("Calculation,HelloWorld,Onboarding", body);
     }
 }
