+++ settings.gradle
@@ -14,5 +14,4 @@ include ':micronaut-camunda-bpm-feature'
 include ':micronaut-camunda-bpm-example'
 include ':micronaut-camunda-bpm-feature:transactional-test-data-jdbc'
 include ':micronaut-camunda-bpm-feature:jetty-webapp-and-rest'
-include ':example-external-task-process'
 include ':example-external-task-worker'
